implementação de permições

+++ app/Http/Controllers/RequestsController.php
@@ -469,7 +469,9 @@ class RequestsController extends Controller
                 if (Auth::user()->hasPermissionTo('delete_request_delivery')) {
                     $buttons .= '<button onclick="return  delete_item_request(\'' . Tools::hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button> ';
                 } else {
-                    $item->status == 2 ? $buttons .= '<button onclick="return  delete_item_request(\'' . Tools::hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button> ' : '';
+                    if (Auth::user()->hasPermissionTo('create_delivery')) {
+                        $item->status == 2 ? $buttons .= '<button onclick="return  delete_item_request(\'' . Tools::hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button> ' : '';
+                    }
                 }
                 if (Auth::user()->hasPermissionTo('create_delivery')) {
                     $buttons .= '<button onclick="return additional_item_request(\'' . Tools::hash($item->product_id, 'encrypt') . '\',\'' . Tools::hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-primary" ><i class="fa-solid fa-pen"></i></button> ';
@@ -481,7 +483,9 @@ class RequestsController extends Controller
                 if (Auth::user()->hasPermissionTo('delete_request')) {
                     $buttons .= '<button onclick="return  delete_item_request(\'' . Tools::hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button> ';
                 } else {
-                    $item->status == 2 ? $buttons .= '<button onclick="return  delete_item_request(\'' . Tools::hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button> ' : '';
+                    if (Auth::user()->hasPermissionTo('create_order')) {
+                        $item->status == 2 ? $buttons .= '<button onclick="return  delete_item_request(\'' . Tools::hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-danger m-t-3"><i class="fa-solid fa-trash"></i></button> ' : '';
+                    }
                 }
                 if (Auth::user()->hasPermissionTo('create_order')) {
                     $buttons .= '<button onclick="return additional_item_request(\'' . Tools::hash($item->product_id, 'encrypt') . '\',\'' . Tools::hash($item->id, 'encrypt') . '\')" class="btn btn-sm btn-primary" ><i class="fa-solid fa-pen"></i></button> ';
